Split into Shared and NonShared serializer tests
